// admin/src/api/generated/schemas/adminCurrencyChainAuditRules.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Internal Wallet Admin API
 * Admin API (served via api-gateway)
 * OpenAPI spec version: v1
 */
import type { AdminCurrencyWithdrawAuditRuleItem } from './adminCurrencyWithdrawAuditRuleItem';

export type AdminCurrencyChainAuditRules = {
  chain_code?: string;
  rules?: AdminCurrencyWithdrawAuditRuleItem[];
};