// admin/src/api/generated/schemas/adminCurrencyChainFeeRules.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Internal Wallet Admin API
 * Admin API (served via api-gateway)
 * OpenAPI spec version: v1
 */
import type { AdminCurrencyWithdrawFeeRuleItem } from './adminCurrencyWithdrawFeeRuleItem';

export type AdminCurrencyChainFeeRules = {
  chain_code?: string;
  rules?: AdminCurrencyWithdrawFeeRuleItem[];
};